requirements/base.txt
anyio==3.5.0
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
beautifulsoup4==4.11.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
decorator==5.1.1
dnspython==2.2.1
email-validator==1.2.0
executing==0.8.3
fastapi==0.75.2
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
ipython==8.2.0
jedi==0.18.1
lxml==4.8.0
matplotlib-inline==0.1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.9.0
Pygments==2.12.0
PyJWT==2.3.0
python-dotenv==0.20.0
PyYAML==6.0
requests==2.27.1
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.35
stack-data==0.2.0
starlette==0.17.1
toml==0.10.2
traitlets==5.1.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
wcwidth==0.2.5
websockets==10.3
